remove legacy Go version compatibility code
