+++ type_alias.go
@@ -1,19 +1,15 @@
-//go:build !go1.9
-// +build !go1.9
-
 package main
 
-import (
-	"go/ast"
-)
+import "go/ast"
 
 func typeAliasSpec(name string, typ ast.Expr) *ast.TypeSpec {
 	return &ast.TypeSpec{
-		Name: ast.NewIdent(name),
-		Type: typ,
+		Name:   ast.NewIdent(name),
+		Assign: 1,
+		Type:   typ,
 	}
 }
 
 func isAliasTypeSpec(t *ast.TypeSpec) bool {
-	return false
+	return t.Assign != 0
 }
