--- type_alias_build_hack_19.go
@@ -1,20 +1,0 @@
-//go:build go1.9
-// +build go1.9
-
-package main
-
-import (
-	"go/ast"
-)
-
-func typeAliasSpec(name string, typ ast.Expr) *ast.TypeSpec {
-	return &ast.TypeSpec{
-		Name:   ast.NewIdent(name),
-		Assign: 1,
-		Type:   typ,
-	}
-}
-
-func isAliasTypeSpec(t *ast.TypeSpec) bool {
-	return t.Assign != 0
-}
