+++ unsafe.go
@@ -1,6 +1,3 @@
-//go:build go1.17 && !go1.20
-// +build go1.17,!go1.20
-
 package main
 
 var g_builtin_unsafe_package = []byte(`
@@ -12,6 +9,9 @@ package unsafe
 	func @"".Sizeof(x ArbitraryType) uintptr
 	type @"".Pointer *ArbitraryType
 	func @"".Slice(ptr *ArbitraryType, len IntegerType) []ArbitraryType
+	func @"".SliceData(slice []ArbitraryType) *ArbitraryType
 	func @"".Add(ptr Pointer, len IntegerType) Pointer
+	func @"".String(ptr *byte, len IntegerType) string
+	func @"".StringData(str string) *byte
 $$
 `)
